requirements.txt
asgiref==3.8.1
Django==5.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.1
django-filter==23.5
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-notifications-hq==1.8.3
reportlab==3.6.12
